Fixed issue when user paused at last word
Now, if the user pauses at last word, when he returns to the document it will be at the beginning.

Also noted another bug, which does not return to the beginning of the document after the text is finished and uses presses play. Subimitted as an bug issue on Zenhub